% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedstat_indicator_info.R
\name{fedstat_indicator_info}
\alias{fedstat_indicator_info}
\title{Download indicator information}
\usage{
fedstat_indicator_info()
}
\value{
data.frame
}
\description{
Download indicator information from https://www.fedstat.ru/organizations/ \cr
Result table contains fedstat indicator id which is needed to request fedstat data \cr
Indicator with condition \code{hidden == TRUE} shows disabled records in fedstat hence ones might not be requested \cr
}
\examples{
\dontrun{
Get all indicator info
get_indicators()
}
}
\seealso{
\code{\link{fedstat_get_data_ids}}
}
