% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedstat_java_script_data_ids_parse_to_json.R
\name{fedstat_java_script_data_ids_parse_to_json}
\alias{fedstat_java_script_data_ids_parse_to_json}
\title{Get data ids from java script source}
\usage{
fedstat_java_script_data_ids_parse_to_json(java_script_source_code)
}
\arguments{
\item{java_script_source_code}{character, java script source code with data ids}
}
\value{
json in form of list with data ids
}
\description{
Get data ids from java script source
}
