% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/fedstat_java_script_default_data_ids_object_ids_parse_to_json.R
\name{fedstat_java_script_default_data_ids_object_ids_parse_to_json}
\alias{fedstat_java_script_default_data_ids_object_ids_parse_to_json}
\title{Get default data ids object ids from java script source}
\usage{
fedstat_java_script_default_data_ids_object_ids_parse_to_json(
  java_script_source_code
)
}
\arguments{
\item{java_script_source_code}{character, java script source code with data ids
and default object ids in it}
}
\value{
json in form of list with 3 character vectors for
lineObjectIds, columnObjectIds, filterObjectIds, which consist of \code{filters_id}
}
\description{
Get default data ids object ids from java script source
}
