% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedstat_parse_all_indicators_names.R
\name{fedstat_parse_all_indicators_names}
\alias{fedstat_parse_all_indicators_names}
\title{Parses https://fedstat.ru/organizations/ indicators hierarchical list into data.frame for reference}
\usage{
fedstat_parse_all_indicators_names(
  url = "https://fedstat.ru/organizations/",
  ...
)
}
\arguments{
\item{url}{https://fedstat.ru/organizations/}

\item{...}{passed to httr::GET}
}
\value{
data.frame, database of all indicators on fedstat.ru with hierarchical structure
}
\description{
Parses https://fedstat.ru/organizations/ indicators hierarchical list into data.frame for reference
}
