% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_js1.R
\name{parse_js1}
\alias{parse_js1}
\title{Get data ids from java script source}
\usage{
parse_js1(script)
}
\arguments{
\item{script}{character, java script source code with data ids}
}
\value{
json in form of list with data ids
}
\description{
Get data ids from java script source
}
