% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_js2.R
\name{parse_js2}
\alias{parse_js2}
\title{Get default data ids object ids from java script source}
\usage{
parse_js2(script)
}
\arguments{
\item{script}{character, java script source code with data ids
and default object ids in it}
}
\value{
json in form of list with 3 character vectors for
lineObjectIds, columnObjectIds, filterObjectIds, which consist of \code{filters_id}
}
\description{
Get default data ids object ids from java script source
}
