% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedstatAPIr-package.R, R/fedstatAPIr.R
\docType{package}
\name{fedstatAPIr-package}
\alias{fedstatAPIr}
\alias{fedstatAPIr-package}
\title{fedstatAPIr: Unofficial API for Fedstat (Rosstat EMISS System) for Automatic and Efficient Data Queries}
\description{
An API for automatic data queries to the fedstat <https://www.fedstat.ru>, using a small set of functions with a common interface.

An API for automatic data queries to the fedstat <https://www.fedstat.ru>, using a small set of functions with a common interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DenchPokepon/fedstatAPIr}
  \item Report bugs at \url{https://github.com/DenchPokepon/fedstatAPIr/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/DenchPokepon/fedstatAPIr}
  \item Report bugs at \url{https://github.com/DenchPokepon/fedstatAPIr/issues}
}

}
\author{
\strong{Maintainer}: Denis Krylov \email{deniskrylovvit@gmail.com}

Authors:
\itemize{
  \item Dmitry Kibalnikov \email{d.kibalnikov@gmail.com}
}


\strong{Maintainer}: Denis Krylov \email{deniskrylovvit@gmail.com}

Authors:
\itemize{
  \item Dmitry Kibalnikov \email{d.kibalnikov@gmail.com}
}

}
\keyword{internal}
