% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedstatAPIr.R
\docType{package}
\name{fedstatAPIr-package}
\alias{fedstatAPIr}
\alias{fedstatAPIr-package}
\title{Unofficial API for fedstat (EMISS) for automatic and efficient data queries}
\description{
This package makes it fairly easy to automatically download filtered data from fedstat.ru, 
    using a small set of functions with a common interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DenchPokepon/fedstatAPIr}
  \item Report bugs at \url{https://github.com/DenchPokepon/fedstatAPIr/issues}
}

}
\author{
\strong{Maintainer}: Denis Krylov \email{deniskrylovvit@gmail.com}

}
